% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumbasehaz.R
\name{cbh}
\alias{cbh}
\title{Internal Function}
\usage{
cbh(lp, event.time, status)
}
\arguments{
\item{lp}{The linear predictor to be used as offset}

\item{event.time}{The event times}

\item{status}{Status indicating the complement of censoring}
}
\description{
Undocumented function. Do not use directly
}
