% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{find_family}
\alias{find_family}
\title{Internal Function}
\usage{
find_family(fam)
}
\arguments{
\item{fam}{Family in character form}
}
\description{
Undocumented function. Do not use directly
}
