% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{formula_setup}
\alias{formula_setup}
\title{Internal Function}
\usage{
formula_setup(
  formula = NULL,
  response.name = NULL,
  linear.name = NULL,
  smooth.name = NULL,
  family = NULL,
  smooth.penalty = NULL,
  num.knots = NULL,
  offset.name = NULL,
  interactions = F
)
}
\arguments{
\item{formula}{A formula to be parsed}

\item{response.name}{The name of the response variable. Vector of two if
\code{family = "binomial"}}

\item{linear.name}{The names of the variables to be used as linear predictors}

\item{smooth.name}{The names of the variables to be used as smoothers}

\item{family}{The family describing the error distribution and link function
to be used in the model. A character string which can only be
\code{"gaussian"} (default), \code{"binomial"}, \code{"poisson"} or
\code{"cox"}. For \code{family = "binomial"}, \code{response} can be
a vector of two and for \code{family="cox"}, \code{weights} must
be provided (see details below).}

\item{smooth.penalty}{The penalty used on the smoothers. Can be 1 or 2}

\item{num.knots}{Number of knots for each smoothers. Can be a single integer
(recycled for each smoother variable) or a vector of integers the same length
as the number of smoothers.}

\item{offset.name}{The name of the offset variable. \code{NULL} (default) if not provided}

\item{interactions}{logical. Should interactions be included.}
}
\description{
Undocumented function. Do not use directly
}
