% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{meandist}
\alias{meandist}
\title{Internal Function}
\usage{
meandist(x, y)
}
\arguments{
\item{x, y}{Vectors of the same length}
}
\description{
Undocumented function. Do not use directly
}
