% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{nzeros}
\alias{nzeros}
\title{Internal Function}
\usage{
nzeros(x, y = NULL)
}
\arguments{
\item{x, y}{Vectors of the same length}
}
\description{
Undocumented function. Do not use directly
}
