% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.gamlasso}
\alias{print.gamlasso}
\title{Print a gamlasso object}
\usage{
\method{print}{gamlasso}(x, ...)
}
\arguments{
\item{x}{fitted model object of the class \code{gamlasso} as
produced by \code{gamlasso}}

\item{...}{Other arguments}
}
\description{
The default print method for a \code{gamlasso} object
}
\details{
Outputs a list of two. \code{lasso} prints the lasso
  model (the same output as \code{print(object$cv.glmnet$glmnet.fit)}) if
  it is non-null and \code{gam} prints the gam model (the same output
  as \code{print(object$gam)}) if it is non-null.
}
\examples{
## Please see the examples in ?gamlasso
}
\seealso{
\code{\link{gamlasso}}, \code{\link{summary.gamlasso}},
  \code{\link[mgcv]{print.gam}}, \code{\link[glmnet]{print.glmnet}}.
}
