% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData}
\alias{simData}
\title{Simulated dataset to be used for gamlasso}
\format{A 100-by-23 data frame. There are 10 variables
  (\code{x1},...,\code{x10}) corresponding to the linear predictors and 4
  (\code{z1},...,\code{z4}) corresponding to the smooth predictors. There are
  7 response variables corresponding to the different models fitted -
  \itemize{
    \item \code{Yg} for the Gaussian response
    \item \code{Yb} as Bernoulli and \code{success} and \code{failure} as
       Binomial count responses
    \item \code{Yp} as the Poisson response
    \item \code{time} and \code{status} as the survival model responses
  }
  The variables starting with \code{X} are the same as the linear predictors but are
  concatenated into a matrix \code{X} to be used for the formula implementation of
  \code{gamlasso}}
\usage{
data(simData)
}
\description{
The package includes a simulated dataset that we will use for the examples.
}
\details{
The code for creating this simulated dataset is included in the
  vignette of this package.
}
\examples{
## Please see examples in ?gamlasso
}
\keyword{datasets}
